Fixed that issue that causes baked imposter shaking on iOS device.
Fixed by changing the accuracy from half to float.

+++ Assets/Plugins/IMP/Shaders/ImposterUnityBillboard.shader
@@ -111,8 +111,8 @@
                 imp.frame2 = IN.plane2;
                 
                 //perform texture sampling
-                half4 baseTex;
-                half4 normalTex;
+                float4 baseTex;
+                float4 normalTex;
                 
                 ImposterSample(imp, baseTex, normalTex );
